LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicy-createXXX-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: TrustedHTML

Harness Error (FAIL), message = ReferenceError: Can't find variable: TrustedHTML

FAIL calling undefined callbacks throws Can't find variable: trustedTypes
FAIL Attributes without type constraints will work as before. Can't find variable: trustedTypes
FAIL trustedTypes.createPolicy(.., null) creates empty policy. Can't find variable: trustedTypes
